카메라 적용 안되고 큐브 띄우기. 리소스 해제 제대로 처리 안함.

+++ FloaterRendererDX11/DX11Mesh.h
@@ -45,7 +45,7 @@ namespace flt
 		Resource<DX11PixelShader> pixelShader;
 
 		ID3D11Buffer* vertexBuffer;
-		size_t singleVertexSize;
+		UINT singleVertexSize;
 
 		ID3D11Buffer* indexBuffer;
 		size_t indexCount;
